+++ client/src/pages/index.js
@@ -2,10 +2,12 @@ import Home from './Home';
 import SendEmail from './SendEmail';
 import Signup from './SignUp';
 import Login from './Login';
+import Dashboard from './Dashboard' 
 
 export {
   Home,
   SendEmail,
   Signup,
   Login,
+  Dashboard
 }
